🐞 fix(unet): fix variable with incorrect type
When generating a new handwriting, at some point the type of the variable x changes from float32 to float64, which causes a RuntimeError for Conv2D. It expected the input data to be of the same type as the bias.

+++ models/LatentDiffusion/unet.py
@@ -21,7 +21,8 @@ class TimestepEmbedSequential(nn.Sequential):
             elif isinstance(layer, SpatialTransformer):
                 x = layer(x, context=context)
             else:
-                x = layer(x)
+                # TODO: Hack - find what causes this to happen
+                x = layer(x.float())
 
         return x
